+++ CHANGELOG.md
@@ -21,7 +21,7 @@ to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 - New spectral source types: Blackbody, LineList and ConstantPhotons
 - New spectrograph
   model: [AtmosphericDispersion](https://stuermer.gitlab.io/pyechelle/example_direct.html#atmospheric-dispersion)
-- support for python 3.11
+- added HDFBuilder, a class to generate .hdf model files from any spectrograph class/model
 
 ## [0.3.5] - 2022-11-30